Added the vao/test.lisp file into the system.
Re-added the flow of the error format string and rest args through
the condition system. Changed the gloss-message messages to reflect
that they are now proper format strings.

Readded kind to the condition, but aren't printing it out yet.

Noticed that %analyze-attribute-spec did all of it work on the spec
itself, so moved it to be before the creation of the attribute-set
hash table. This allowed me to move the duplicate check to there too
and is cleaner.

Changed the condition strings to hold the full attribute specs that
failed whatever constraint being checked. This shows you what values
everything had for faster debugging.